separates default cash flow generation from custom
